fix: del propertySource in AppConfig, refactor
