+++ student-test-app-part3/src/main/java/ru/otus/config/AppConfig.java
@@ -3,14 +3,12 @@ package ru.otus.config;
 
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
-import org.springframework.context.annotation.PropertySource;
 import org.springframework.context.support.ConversionServiceFactoryBean;
 import org.springframework.core.convert.converter.Converter;
 
 import java.util.Set;
 
 @Configuration
-@PropertySource("classpath:app.yml")
 public class AppConfig {
 
     @Bean
