+++ student-test-app-part3/src/main/java/ru/otus/service/MessageService.java
@@ -1,0 +1,7 @@
+package ru.otus.service;
+
+public interface MessageService {
+    String getMessage(String message, String... args);
+
+    String getMessage(String message);
+}
