+++ student-test-app-part3/src/main/java/ru/otus/service/MessageSourceService.java
@@ -1,7 +1,24 @@
 package ru.otus.service;
 
-public interface MessageSourceService {
-    String getMessage(String message, String[] args);
+import lombok.RequiredArgsConstructor;
+import org.springframework.context.MessageSource;
+import org.springframework.stereotype.Service;
 
-    String getMessage(String message);
+@Service
+@RequiredArgsConstructor
+public class MessageSourceService implements MessageService {
+
+    private final LocaleProvider provider;
+
+    private final MessageSource messageSource;
+
+    @Override
+    public String getMessage(String message, String... args) {
+        return messageSource.getMessage(message, args, provider.getLocale());
+    }
+
+    @Override
+    public String getMessage(String message) {
+        return messageSource.getMessage(message, null, provider.getLocale());
+    }
 }
