+++ student-test-app-part3/src/test/java/ru/otus/service/ApplicationRunnerTest.java
@@ -21,7 +21,7 @@ class ApplicationRunnerTest {
     private TestingService testingService;
 
     @Mock
-    private MessageSourceService messageService;
+    private MessageService messageService;
 
     @Mock
     private IOService ioService;
